agent/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-silero>=0.6.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
